requirements.txt
flask==3.0.0
smbus2==0.4.2
requests==2.31.0
Adafruit-GPIO==1.0.3
RPi.GPIO==0.7.1